Add typemaps for `DrawPoly` functions
These functions in Raylib accept an array of structs, represented by a
pointer to first element of the array and a count. This type of direct
pointer referencing is not possible in Ruby, so we need a SWIG typemap
to allow the Ruby wrapper function to accept a Ruby array.