Create weather display showing random data
The weather display has all the functionality needed to display the
weather. However, the weather is not actually retrieved yet, so random
data is shown. Errors are randomly introduced to exercise the display of
error states.

Currently, only the new weather display component is being shown. In the
future, the application will automatically cycle between the time and
weather displays.