Update to latest raylib
The specific change is a change to the enum name `BUTTON_BUTTON_LEFT`.
This change was made in [this commit to Raylib](https://github.com/raysan5/raylib/commit/2545f62565fd246d1c59bf9a6bcf4942f4ad12ad),
which at the time of writing is not part of an official Raylib release.

The change does define a pre-processor symbol for backwards
compatibility, but SWIG doesn't automatically pick that up. In any case,
the future is the new name, so I switched to it.